add comment for all files

+++ run_forker.sh
@@ -1,6 +1,9 @@
 #!/bin/bash
 timestamp=`date "+%m%d_%H%M%S"`
 
-rm result.csv
+if [ -e result.csv ] 
+then
+    rm result.csv
+fi
 make forker_run
-python3 util/visual_forker.py -t $timestamp
+python3 util/visual_forker.py -t $timestamp -d
