+++ run_vanilla.sh
@@ -1,6 +1,9 @@
 #!/bin/bash
 timestamp=`date "+%m%d_%H%M%S"`
 
-rm result.csv
+if [ -e result.csv ] 
+then
+    rm result.csv
+fi
 make vanilla_run
 python3 util/visual_vanilla.py -t $timestamp -d
